requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
openai>=1.0.0
ezdxf>=1.0.0
fpdf==1.7.2
numpy>=1.21.0
matplotlib==3.7.1
requests>=2.26.0
beautifulsoup4==4.12.0
pandas>=1.3.0
python-dotenv>=0.19.0
openpyxl>=3.0.0
PyMuPDF>=1.20.0
Pillow>=9.0.0
tqdm>=4.62.0 